requirement.txt
langchain
langchain-core
langchain-community
langgraph
langchain-mistralai
langchain-openai
langchain-groq
langchain-google-genai
python-dotenv
faiss-cpu
pypdf
tiktoken
unicorn
fastapi
requests
streamlit==1.39.0
langchain-text-splitters==0.0.1
chromadb==0.4.24
pydantic==2.6.4

